requirements.txt
# Django and related packages
Django==5.2.5
djangorestframework==3.14.0
channels==4.3.1
daphne==4.2.1
asgiref==3.9.1
sqlparse==0.5.3
tzdata==2025.2

# Database adapter
psycopg2-binary==2.9.10

# API access and networking
requests==2.32.3
urllib3==2.2.1

# Google Gemini AI API
google-generativeai==0.3.1
google-api-core==2.18.0
google-auth==2.28.1
googleapis-common-protos==1.63.0
protobuf==4.25.3

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0
Pillow==10.2.0

# Configuration and environment management
python-dotenv==1.1.1

# Web socket and async support
autobahn==24.4.2
Twisted==25.5.0
txaio==25.6.1

# Security and cryptography
cryptography==45.0.6
pyOpenSSL==25.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
idna==3.10

# Development tools
black==25.1.0
isort==6.0.1

# Utility packages
colorama==0.4.6
tqdm==4.66.2
pytz==2024.1
typing_extensions==4.14.1
six==1.16.0